Move RegisterStep into associated behavior
